+++ src/Modules/Base/Subscriptions/index.ts
@@ -1,12 +1,13 @@
 import ModuleBuilder, { Module } from "../../../types/Module";
 import subscription from "./commands/subscription";
+import inner from "./inner";
 import scheduler from "./scheduler";
-import reminder from "./tasks/reminder";
 
 export default new ModuleBuilder("subscriptions", (module: Module) => {
 	
-	subscription(module) // Command /subscription
-	scheduler(module)    // Tasks initiator
+	subscription(module)	// Command /subscription
+	scheduler(module)		// Tasks initiator
+	inner(module)			// Inner api routes
 
 	return module
 })
